fuzz: add a target for the coins database
Summary:
It reuses the logic from the `coins_view` target, except it uses an
in-memory CCoinsViewDB as the backend.

Note `CCoinsViewDB` will assert the best block hash is never null, so we
slightly modify the coins_view fuzz logic to take care of this.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/32602 | core#32602]]
https://github.com/bitcoin/bitcoin/pull/32602/changes/46e14630f7fe8e2d51adc18a4f50345eb26970c7
https://github.com/bitcoin/bitcoin/pull/32602/changes/cfc42ae5b7ef6d3892907cfe36dc3ae923495d37

The missing commit (*avoid underflow in coins_view target*) is not applicable because it  would be reverted by D20082 if we had done the backports in the correct order

Test Plan: `ninja bitcoin-fuzzers`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D20097