build: swap otool for (llvm-)objdump
Summary:
```
Similar to libtool, (llvm-)otool only exists with a version suffix
on some systems (Ubuntu), which makes it annoying to use/find. Avoid
this, by switching to objdump. Which is a drop-in replacement.

This is related to #21778, and the switchover to using vanilla LLVM for
macOS.
```

Backport of [[https://github.com/bitcoin/bitcoin/pull/29739 | core#29739]].

Depends on D17258.

Test Plan: Run the GUIX osx build

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Subscribers: PiRK

Differential Revision: https://reviews.bitcoinabc.org/D17259