[cashtab-components] Remove cashtab-components from monorepo
Summary:
cashtab-components has not been updated or maintained apart from maintenance releases updating npm packages in years.

To be useful, needs to be rebuilt on chronik. At this point, it would be more efficient to write it from the ground up than to iteratively improve from its current state.

PayButton is doing more to support this use case. If we have a need for this product, should bring it back. At the moment, this is low priority and does not have enough impact to merit a presence in the monorepo.

Test Plan: Any thoughts on keeping it here, confirm no missed references.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D16320