[electrum] rework copyright and license files, link it in About dialog
Summary:
The copyright dates were outdated in the About dialog in the application. Remove them from this dialog and just put a link to a "COPYING" file at the root of the Electrum ABC subrepository. Don't put any date ranges in the copyright notices, as these are unlikely to be maintained on a regular basis. See https://www.linuxfoundation.org/blog/blog/copyright-notices-in-open-source-software-projects

Rename the "LICENSE" file "COPYING" and add the copyright information, to mirror what is done in for the node software.

Test Plan:
Run the application, open the About menu and check that it works.
The URL does not work yet, as the COPYING file is not yet on the github mirror of the monorepo, so test is with this patch
```
diff --git a/electrum/electrumabc_gui/qt/main_window.py b/electrum/electrumabc_gui/qt/main_window.py
index d06b69b646..e51f8025b6 100644
--- a/electrum/electrumabc_gui/qt/main_window.py
+++ b/electrum/electrumabc_gui/qt/main_window.py
@@ -1051,7 +1051,7 @@ class ElectrumWindow(QtWidgets.QMainWindow, MessageBoxMixin, PrintError):
                 f"parts of the {CURRENCY} system."
             )
             + "</p></span>"
-            + f"<p><a href={REPOSITORY_URL}/blob/master/electrum/COPYING>"
+            + f"<p><a href={REPOSITORY_URL}/blob/master/electrum/LICENCE>"
             + _("License and copyright information")
             + "</a></p>",
         )
```

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D16326