[Cashtab] Deprecate (unimplemented) function for explicit slp1 burns
Summary:
Intended to implement this in special cases where the user wants to burn a token amount that matches a utxo amount, which is common-ish (happens for example when user wants to burn all).

However, we don't really get any advantage from this. The legacy way of burning is indexed fine by chronik and is much more flexible (user may burn any amount).

Deprecating here as it is not implemented and, since now there are no plans to implement it, does not make sense to upgrade to ecash-lib style.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D16338