[Cashtab] Remove duplicated test
Summary:
Depends on D16338

We formerly had two tests as we formerly tested on NNG chronik mocks and in-node chronik mocks prior to migrating to in-node chronik-client. Missed cleaning this up in the migration.

Test Plan: `npm test`, can copy paste the omitted one into a diffchecker and see they run the same tests, tho the test names are slightly different (including a burn copypasta name)

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D16339