[Cashtab] Improve target output generation tests
Summary:
Depends on D16339

When tests were written, did not have a good solution to the problem of "cannot store Buffer in mocks." So, tests were designed to fit around imagined restriction of "mocks should only be the hex string."

This led to complicated tests and even some missed cases as functions were improved.

Refactor tests so that we are exactly matching mocks.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D16340