[electrum] introduce a new ECPubkey class
Summary:
and start using it in transaction.py to verify tx signatures.
This makes the `MyVerifyingKey` private to ecc.py.

In a following diff, it will be used to verify other type of messages as well, and will serve as a base class for a ECPrivkey class that will replace the current ECKey class.
These new classes will encapsulate the interaction with the ecdsa lib, so that other modules don't need to know about this implementation detail.

Depends on D16348

This is split out of D16301

Test Plan: `python test_runner.py`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D16349