[Cashtab] Patch typo in css
Summary:
Discovered bug while working on nft listing diff, fix is unrelated to nft diff and should be isolated

There is no expected UI change. This style was already inherited, which is why the typo was never noticed. Makes sense to keep it in this specific section for readability / edit-ability.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D16371