test: Add diamond-shape prioritisetransaction test
Summary:
Backport of [[https://github.com/bitcoin/bitcoin/pull/24623 | core#24623]] and [[https://github.com/bitcoin/bitcoin/pull/24637 | core#24637]] for the wallet tooling.

The diamond test is not really testing the same thing as the source material because we don't track the fees including ancestors/descendants anymore, but it demonstrates that the *_transfer_multi functions are working.

Depends on D16412.

Test Plan:
  ./test/functional/test_runner.py mining_prioritisetransaction

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D16413