[Cashtab] Update paywall article link
Summary:
The current paywall article link in tx history is a placeholder pointing to the tx on explorer where the article was created.

Now that direct sharing of articles has been in use for some time, this diff updates it to point directly to the article itself.

Test Plan: open a wallet that has received or sent a paywall payment, click on `Paywall Article` and ensure it directs to the article on eCashChat.

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D16459