Use DOGE and µDOGE unit insted of BCHA and XEC
Dogecoin has 8 decimal places by default and the DOGE ticker.

To not break all tests, we add the -usedogeunit command line flag, which is similar to -ecash, except it is on by default, and changes the units to 8 places (without affecting the address prefix), and then we disable -usedoge=0 on the test framework.