[explorer] Properly handle html chars in token ticker
Summary:
If a token ticker contains some html chars it can render weirdly on the amount display due to missing escapes. Let's properly escape the chars so the rendering is always correct.
Note that the templated content is not affected, only the amount rendered in the js code needs to be updated.

Test Plan:
Hardcode a token with some <> chars in renderOutput.
Check the explorer shows the amount correctly in the list of txs.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D16521