Get rid of ConnectTrace
Summary:
This class is no longer used for anything useful since the validation interface has split SyncTransaction() into TransactionAddedToMempool(), BlockConncted() and BlockDisconnected(). Getting rid of the legacy class simplifies the code a bit. The only change in behavior is that the block connection signals are sent prior to the transactions added to mempool in case of a reorg, which is more correct anyway.

Depends on D16538.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D16539