test: fix constructor of msg_tx
Summary:
In python, if the default value is a mutable object
its shared over all instances, so that one instance being changed
would affect others to be changed as well.
This was likely the source of various intermittent bugs in the
functional tests.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/30552 | core#30552]]

Test Plan: `ninja check-functional`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18896