[e.cash] Improve anchor links on core tech
Summary:
The anchor links on /core-tech land with the content at the very top of the page. With the
navbar overlay it hides the content. Adding an invisble block to use instead to provide some relief
from the top of the page

Test Plan: Preview the site and try the core tech links on the landing page. They should scroll to a good position

Reviewers: #bitcoin_abc, Kronkmeister

Reviewed By: #bitcoin_abc, Kronkmeister

Differential Revision: https://reviews.bitcoinabc.org/D16599