[Cashtab] Update tx history to parse eCashChat authentication txs
Summary:
v1.5.2 of eCashChat will utilize onchain authentication transactions to verify ownership of addresses on mobile devices. This is done by passing a BIP21 querystring to Cashtab.com on mobile devices with a pre-populated dust send tx.

It's done this way because the use of signature verification is very laborious and Cashtab Extensions is not available on Android/iOS.

This diff updates Cashtab to recognize these transactions.

{F8284667}

Test Plan: npm test

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D16620