[examples and e.cash] Deprecate examples
Summary:
The examples repo showcased some simple scripts for basic chronik methods. It also highlighted tx construction using utxo-lib and ecash-coinselect.

It is now obsolete for a handful of reasons:

- in-node chronik-client instead of NNG
- ecash-lib instead of utxo-lib
- schnorr signed txs of ecash-lib obsoleting ecash-coinselect

I do not think it is worth updating this repo to use the latest libraries, because

- the chronik-client integration tests are more exhaustive, always up to date, and more detailed than these exmaples
- utxo-lib and ecash-coinselect are obsolete now, but we still have some more work to do in getting utxo selection sorted out in ecash-lib

Test Plan: CI test, preview e.cash and confirm the examples repo is no longer linked by the build page

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D16629