test: refactor: Move g_insecure_rand_ctx.Reseed out of the helper that calls MakeRandDeterministicDANGEROUS
Summary:
The global g_insecure_rand_ctx will be removed in the future, so
removing it from this helper is useful.

Also, tying the two concepts of the global internal RNGState and the
global test-only rng context is a bit confusing, because tests can
simply use the m_rng, if it exists. Also, tests may seed more than one
random context, or none at all, or a random context of a different type.

Fix all issues by moving the Reseed call to the two places where it is
used.

This is a partial backport of [[https://github.com/bitcoin/bitcoin/pull/30571 | core#30571]]
https://github.com/bitcoin/bitcoin/pull/30571/commits/fa19af555dff6d6c722caf36319b158699d2aa95
Depends on D19256

Test Plan: `ninja check bitcoin-fuzzers`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D19257