[electrum] grind ecdsa low R signature when using libsecp256k1
Summary:
This is a partial backport of [[https://github.com/spesmilo/electrum/pull/5820 | electrum#5820]]

See also D4467
Note that Electrum ABC produces schnorr signatures for most transactions, so for most users this is not going to change much. It makes future backports a bit easier by adhering more closely to the Electrum code.

And I didn't backport the change in the pure python version of signing as this code is going to be dropped very soon.

Depends on D16684

Test Plan: `python test_runner.py`

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D16685