[avalanche] Make buildRandomProof use a random payout script
Summary: It is easy to accidentally write tests assuming that buildRandomProof() will produce proofs with random payout scripts. This is not the case. There does not appear to be good case for random proofs to payout to UNSPENDABLE_ECREG_PAYOUT_SCRIPT considering all tests pass without any modification.

Test Plan:
  ninja check-avalanche

Reviewers: O1 Bitcoin ABC, #bitcoin_abc, PiRK

Reviewed By: O1 Bitcoin ABC, #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D16691