[electrum] misc fixes for OSX build
Summary:
This fixes a couple of non-fatal errors when running the make_osx script:
 - pyinstaller 6.4.0 requires pyinstaller-hooks-contrib>=2024.0
 - readlink: illegal option -- f

The make_osx script already uses `greadlink -f`, so we can assume that it will be available for the `base.sh` script as well.

Depends on D16718

Test Plan:
`contrib/osx/make_osx`

Look at the entire log for errors.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D16719