[chronik.e.cash] warn on broken links during docusaurus build
Summary:
Broken link errors are causing this build to fail in CI

The links are not actually broken. I am not sure why they are being detected as such. We have this same problem with the markdown linter, which is why `arclint` is ignoring this repo.

It's an issue that should be resolved (if practical) -- at the moment tho, looking to fix the broader npm module problem.

Test Plan:
```
docker build -f chronik.e.cash.Dockerfile -t chronik-docs .
```
builds

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D16720