[Cashtab] Do not lock UI while token info is loading on NFT screen
Summary:
A number of debottleneck opportunities are available to improve the load time of the NFT screen.

This is the easiest one.

We do not actually need the token info to render all the listings. Render the listings when you get them with a spinner for info that comes from token cache.

We can also dramatically improve the speed of getting and caching this token info, but that should be another diff.

Test Plan:
`npm test`, create a wallet and go to the "Listed NFTs" screen on this test site, check the console log, you can see how slow the token info is loading and that it does load (and when it does, we see token info).

https://66e6bed8a1ec4780d1e3866b--cashtab-local-dev.netlify.app/

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D16763