[e.cash] Update core tech section on landing page
Summary:
Some people have pointed out the hands on the core tech section are a bit pixelated and they
dont really relate to the content much. The section also presents a wall of text to the user.

This is a proposed design overall to the section to make the information easier to digest at a glance
and hopefully just look a bit cooler

Test Plan: preview the site and check it out

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Subscribers: bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D16768