[e.cash] Improve copy in Core-Tech segment on home
Summary:
Removed redundant sentences, made explanations less ambiguous, easier to understand, and kept things birds-eye.
This also makes "Learn More" pages less redundant as you now actually learn more when going there, allowing for the core-tech segment to be more concise.

Also changed "eTokens and NFTs" back to eTokens. We did this before to have NFTs be visible. But it makes little sense to have it this way and makes the use of "eTokens" ambiguous, which doesn't fit to the rest of the copy talking about "tokens" instead. Since it is about Core-Tech, we should use "eTokens" referencing the protocol, while "tokens" and "NFTs" should be the things you can do with said protocol. It looks cleaner on the page as well. NFTs are mentioned in the explainer, which will have to suffice.

Test Plan: Visit e.cash, scroll down to Core-Tech segment.

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Subscribers: bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D16799