[e.cash] Prevent image jump on /get-ecash
Summary:
If you scroll to the bottom of /get-ecash, the animated images in the mining and staking sections are jumping over after a moment.
This is from the fact they are absolutely positoned and the section is also animating in. They dont know where to be until the section animation is done
Removing the section animation as it doesnt add much and will avoid this weird jump on the images

Test Plan: preview the site and check you dont see the jump

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Subscribers: bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D16808