[ecash-agora] Add `historicOffers` function to `Agora`
Summary:
Allows querying offers made in the past, which were either cancelled or accepted. Unspent offers are ignored.

This reflects what users would expect an "exchange" UX pretty closely: "activeOffers*" is the orderbook, "historicOffers" are the candlesticks (and also the cancelled offers, but those would have to be ignored).

They can also be queried by pubkey, giving users a history of their offers.

This is done by going through the tx history of the token ID/group token ID/pubkey group of the agora plugin, and looking at *inputs* that spend an agora output. We then reconstruct the `Utxo`, and then we just re-use the `_parseOfferUtxo` function, which we also gave a status.

Test Plan: `npm run integration-tests`

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D16819