[avalanche] Backport heartbeat activation code
Summary:
The upstream change activated heartbeat but we don't do that here. Only
backporting to keep up with codebase changes.
---
As per title. We deactivate the feature by default in tests because it doesn't play nicely with regtest. The feature enable flag is now public and documented.

Release notes will be added later in the version bump diff.

Test Plan:
  ninja all check-extended check-upgrade-activated-extended

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Subscribers: Mengerian

Differential Revision: https://reviews.bitcoinabc.org/D16823