[Iguana] Add `-format=csv` to iguana
Summary:
This allows users to view the script trace in a spreadsheet software.

An example how this would look like in Google Sheets (here a Agora Partial trace):
https://docs.google.com/spreadsheets/d/1KJLV9nhkMzXTI3dzjWXBAuXiUX61od15p1ztscUqOUg/edit?usp=sharing

In the future, we can extend the CSV file with additional columns, like the number of bytes used so far (for the 520 push limit) and the number of non-push ops so far (for the 201 opcode limit).

Depends on D16851.

Test Plan: `ninja check-iguana`

Reviewers: Fabien, #bitcoin_abc

Reviewed By: Fabien, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D16852