[herald] Fix parsing of the staking reward winner
Summary:
The parsing function returns the staker address. It's very possible that there is no address for the winner, so avoid crashing the app if this happens.
This diff lets the parsing return the script hex instead, and the callsite attempt to convert it to address so it will make no difference with the current network state.

Test Plan:
  npm test

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D16861