[Cashtab] Add horizontal padding to buttons
Summary:
We just happen to not have any buttons that use close to most of the width in Cashtab at the moment. This could happen after supporting agora partial trades, as we include token names on buttons.

We should have something besides 0 here anyway, we don't want text that runs right up to the button border.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D16921