[Cashtab] Parse agora ad prep txs in tx history
Summary:
Parse agora ad prep txs in Cashtab tx history

These are currently rendered as SLP SEND txs. They are SLP SEND txs, but a special case where we are sending to a p2sh script that is then immediately used to generate the offer.

We do not see the actual listing tx in Cashtab for SLP 1 agora listings, because we are not getting the tx history of the generated p2sh script. NBD, since in Cashtab we know this tx is always creating a listing.

Will see if can add offer details (price etc) in a reasonable way in a later diff. For now, at least parse correctly.

Note this also works for Agora Partials.

Test Plan:
`npm test`, check screenshots

listing an nft:
{F9789951}

listing slp token:
{F9789880}

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D16960