[e.cash] Change "Stack" to Stack Wallet and add "Desktop"
Summary: There was a request to add "wallet" to the stack wallet name and mention the desktop versions. Also added "wallet" to "Unstoppable" while at it. Reason is to make it more specific and easier to find when searching for it online.

Test Plan: go to e.cash/wallets check Stack wallet copy

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D16988