[Cashtab] Fix agora depth bar fill effect
Summary:
Lost this when converted all the math to bigints

In this case, we need to use either Number or a BigNumber library. (smaller bigint) / (bigger bigint) will always be 0 ... until in Agora, it can be 1, when acceptTokens = offeredTokens.

Since this value is only used for shading UI, we do not need great precision, and `Number` is fine.

Test Plan:
`npm test`, review screenshots

current behavior: we do not show accepted portion on depth bars until we are accepting 100% of it
{F10150619}

after: expected behavior is back (the portion of the candle the user is accepting with the slider is highlighted in green)
{F10150613}

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D17016