[ecash-herald] Parse agora actions
Summary:
Summarize SLP1 agora partial txs in daily summary

We now use ecash-lib and ecash-agora in ecash-herald, so update CI settings and herald dockerfile

Arguably ecash-agora does not need to be imported to the herald yet, we could do the agora parsing without it. However, we will for sure need it later if we want to add volume metrics, which we do. So might as well add it and update the CI stuff now.

Test Plan:
`npm test`, see screenshot

`docker build -f ecash-herald.Dockerfile -t ecash-herald_local .` for new docker file

CI tests to confirm `ecash-agora` change does not impact the lib or other apps using it

last 24 hrs:
{F10164586}
{F10164587}
{F10164589}

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D17019