[Cashtab] Use ABC chronik first
Summary:
We have been using pay2stay server as it (previously) was the only known instance to have chronik offers indexed.

chronik-native1 now has this. So, default to this server first. Fallback to pay2stay on an error. Fallback to chronik-native2 if both are in error (in this case, agora and NFTs will not work).

Test Plan: `npm test`

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D17042