[ecash-herald] Patch axios syntax bug in getNextStakingReward
Summary: Axios reply returns expected shape in `data`. We check this, but then get {address, scriptHex} from the raw response and not the data --- so we can never get it.

Test Plan: I created a script and checked the console.log to confirm. We could test this function, but we need to test the mock cache and some other things. I found this while converting herald to typescript, imo the test can wait for that conversion.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D17057