[Cashtab] Add canvas mock to prevent annoying jest debug logs
Summary:
Most cashtab integration tests get this logging because canvas is not mocked in jest environment. It does not need to be for any test reason except that the logs are distracting and make it harder to find real problems.

Add the mock to suppress these logs.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D17083