[Cashtab] Update and implement jest setup file
Summary:
In the last diff where I made a jest setup file, I forgot to implement it (see `jest.config.js` change here).

It seemed to be working because we no longer needed all the Window prep work for tests to pass. But, this must be due to using a newer version of jest or react testing library, as the file written in that diff was not actually implemented. So I just removed those.

Here, we implement jest.setup.js, and remove the obsolete Window statements. We extend global mocks for bip39.

We are not able to mock recaptchav3 globally (used in Home and Rewards) because it requires a constant used in those tests.

Test Plan:
`npm test`

`grep -r "window.sideshift" src/` and no output from test mocks
`grep -r "jest.mock('bip39'" src/` and no output

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D17093