[Cashtab] Support minting additional NFT Parent tokens
Summary:
Depends on D17091

We are currently rendering this method for NFT1 Parent tokens with mint batons, but it fails because we do not specify type.

It's a bug / oversight. Patch by passing the token protocol number, which is available from chronik utxo and is required input for the ecash-lib slp method.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D17097