[chronik] implement scripthash type for utxos endpoints
Summary:
This is a follow up to D16937 adding a scripthash option to the utxos script endpoint.
We also look for the scripthash in the mempool index added in D17037 in case this script didn't make it into the db yet.

This one was a bit trickier because we need to provide the script, which was previously built from the query itself. Now if the query passes a scripthash payload, we need to find it in the db and decompress it.

Depends on D17037

Test Plan: `ninja check-functional`

Reviewers: #bitcoin_abc, tobias_ruck

Reviewed By: #bitcoin_abc, tobias_ruck

Subscribers: tobias_ruck

Differential Revision: https://reviews.bitcoinabc.org/D17043