[Cashtab] Modify csp headers to allow base64 font used in swiper
Summary:
Getting swiper css working with `npm start` and also on netlify test deployments. But not on Cashtab prod. Looking into network, the font file fails to load due to CSP restrictions.

Add CSP directive in header to allow loading base-64 font.

Test Plan: Hard to test this without seeing it with the header, which requires the URL. So, will see when it lands.

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D17127