[avalanche] Return pending stake contender status when there is no local staking rewards computed
Summary:
Before staking rewards is computed on the local node, we cannot be sure
of our vote on any particular contender. To avoid convergence issues, we
respond with a pending status.

This also makes contender voting slightly more robust in the unexpected event
that staking rewards fails to be computed for a block due to a bug or
otherwise.

Test Plan:
  ninja check-avalanche

Reviewers: O1 Bitcoin ABC, #bitcoin_abc, Fabien

Reviewed By: O1 Bitcoin ABC, #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D17144