[DogeCash] Fix feature_legacy_script_flags.py
Fix the test so it tests the correct script flag (and sighash type) behavior.

This test was written under the assumption that there are more mandatory script flags on Dogecoin, however, the only one is SCRIPT_VERIFY_P2SH.

Somehow this test was merged (probably by having the wrong script flags first, then implementing the test, and later fixing the script flags without fixing the test), so we fix it here.