[DogeCash] Add `CalcExpectedMerkleTreeIndex` (#48)
This function allows tweaking the indices at which the hash of our block occurs in the merkle tree in the merge-mined block.

The design of this seems a bit obscure, and the merge mine nonce is 0 for most blocks on the network, but it's a consensus rule so we'll have to follow it exactly.