[DogeCash] Disable avalanche until it is ready (#57)
Avalanche per-se is not an issue, however, if it's enabled, it disables deep reorgs. This means that by default, nodes would get stuck on a shorter chain on a deep reorg.

The simplest way to fix this is to disable Avalanche by default until we're positive Dogecoin will activate it.

Co-authored-by: Tobias Ruck <ruck.tobias@gmail.com>