[ecash-herald] Add staker info to block msgs
Summary:
Fetch data from CoinDance api to get information about active stakers. Include the staking rwd winners chances of winning, and total stake (if you fetch successfully from the API).

Now that we need `secrets` for this API key and tests, generate this file if it is not in place to keep CI working. Same model used for token-server.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D17180