[cashtab] UI Overhaul
Summary:
As discussed here are some proposed UI changes to the app as a whole. It is maybe more changes than the
typical diff, but it kind of needs to be done in one go. It is all mostly CSS changes in any case.

A new desktop layout and color changes are the main updates here, with more granual design improvements to individual pages to come

Test Plan: preview the site and check it looks good on mobile and desktop

Reviewers: #bitcoin_abc, bytesofman, emack

Reviewed By: #bitcoin_abc, bytesofman, emack

Subscribers: emack, bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D17186