[avalanche] Fix getdata log not showing stake contender inv type name
Summary: Turns out the "received getdata" log line only spits out hex if the inv type does not have an associated case in CInv::GetCommand(). This should be fixed for stake contenders because the hex is not only confusing but makes it more difficult to write tests for multiple inv types.

Test Plan:
  ninja check check-functional

Reviewers: O1 Bitcoin ABC, #bitcoin_abc, PiRK

Reviewed By: O1 Bitcoin ABC, #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D17189