kernel: Move default cache constants to caches
Summary:
They are not related to the txdb, so a better place for them is the
new kernel and node cache file. Re-use the default amount of kernel
cache for the default node cache.

This is a partial backport of [[https://github.com/bitcoin/bitcoin/pull/31483 | core#31483]]
https://github.com/bitcoin/bitcoin/pull/31483/commits/65cde3621dbb9ac7d210d4926e7601c4adf5f498

Depends on D18582

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18583