[bip21] Introduce a spec for supporting simple token send txs
Summary:
This would make it easy for users to share token payment requests via QR code or cashtab url

While this is already possible to do using existing bip21 `op_return_raw`, it is not straightforward for non-developers to create such a request.

Test Plan: Proofread. To be implemented in Cashtab.

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D17196