[Dogecash] Fix `TRACK_SIZE_VERSION`, update version to 0.2.0
Currently, we don't track the size of blocks in `CDiskBlockIndex`, because we didn't update the `TRACK_SIZE_VERSION`. This means people are prompted to upgrade their DB at every version bump.

This fix unfortunately requires everyone to -reindex their DB (which is unavoidable since the data isn't there), but at least this time they only have to do it once.