[mock-chronik-client] Use npm published version of ecashaddrjs as dependency for published version
Summary: The local dependency of `ecashaddrjs` is great for the monorepo as it ensures we always use the latest version in CI. However this breaks the published version. Since `ecashaddrjs` is a `dependency` of `mock-chronik-client` and not a `devDependency`, it's actually not even possible to `npm i mock-chronik-client` right now. The published version is broken.

Test Plan:
`docker build -f mock-chronik-client.Dockerfile -t mock-chronik-client_local .
docker run mock-chronik-client_local` builds

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D17227