[mock-chronik-client] Add new dummy method to allow Agora construction with MockChronikClient
Summary:
We add this method in Cashtab tests. Would be nice if it were handled in the lib.

We can't test this here without adding the whole `ecash-agora` dep to this lib, which is imo not worth it.

Verify that it doesn't break this with current tests passing.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D17279