Add a new getinfo rpc
Summary:
This returns the node version (which currently is in getnetworkinfo, which is counter intuitive) and whether avalanche is enabled.
This gives a fast (no lock) method to get the version and whether the node is avalanche enabled, to be used in the functional tests in a follow-up.
This RPC will be much faster than the bloated getnetworkinfo for such simple data.

Test Plan:
  ./test/functional/test_runner.py abc_rpc_getinfo

Reviewers: #bitcoin_abc, tobias_ruck, PiRK

Reviewed By: #bitcoin_abc, tobias_ruck, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D17282