[test] Only sync proofs when avalanche is enabled
Summary:
This makes it possible to use the functional test framework with avalanche disabled nodes.
Since the nodes are synced by default upon startup, all the tests currently would fail without this change except abc_rpc_getinfo. The flag order reversal in this test would cause it to fail as well on master but it succeeds after this patch.

Depends on D17282.

Test Plan:
  ./test/functional/test_runner.py abc_rpc_getinfo

Reviewers: #bitcoin_abc, PiRK, tobias_ruck

Reviewed By: #bitcoin_abc, PiRK, tobias_ruck

Subscribers: tobias_ruck

Differential Revision: https://reviews.bitcoinabc.org/D17283