[Cashtab] Remove swap screen
Summary:
Network analysis shows (some, not all) speed issues are sideshift-related. We are not really getting benefits from this integration. Need to do an API integration when we have a more compelling use case, like selling etokens to ETH users.

Anyway, the widget is not working now. It is easy enough to bring it back.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D17291