Revert "Revert "[Cashtab] Convert wallet context to ts""
Summary:
This reverts commit b88a433009eaca5a4fcb685f7a611665f115881b.

Identified and resolved the bug that caused this to fail builds in prod. Was an issue with not properly `initWasm()`

See updated file `cashtab/src/index.tsx` and the much lighter lint-related changes that are applied here vs in the original diff.

Test Plan:
`npm test`
`npm start` and app loads
`npm run build` and app builds

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D17311