[chronik] add a test for the new JSON RPC interface
Summary:
Add a basic functional test for the new Chronik Electrum interface added in D17185.
This adds a few methods to the test framework that can be reused for further testing of JSON RPC interface as more methods are added.

Reduce the port range again to accomodate for this third family of ports (see D16048)

Ref T3598

Test Plan: `ninja check-functional`

Reviewers: #bitcoin_abc, tobias_ruck, Fabien

Reviewed By: #bitcoin_abc, tobias_ruck, Fabien

Subscribers: Fabien, tobias_ruck

Maniphest Tasks: T3598

Differential Revision: https://reviews.bitcoinabc.org/D17314